+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Fluentify will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased]
+
+## Added
+
+- Introduced analyzer `FLTFY08` to warn when records annotated with `Fluentify` are not partial and need a generated parameterless constructor.
+
 # 1.8.2 - 2025-11-22
 
 ## Fixed